requirements.txt
celery==5.1.0
Django==3.0.4
django-cors-headers==3.4.0
django-extensions==3.1.1
django-redis==4.12.1
django-rest-swagger==2.2.0
django-sslserver==0.22
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
docx2pdf==0.1.7
httplib2==0.9.2
idna==2.6
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
moviepy==1.0.3
mysql-connector-python @ http://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.zip
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
openapi-codec==1.3.2
pandas==1.1.5
paramiko==2.0.0
pexpect==4.2.1
Pillow==5.1.0
proglog==0.1.9
prompt-toolkit==3.0.18
protobuf==3.0.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyfcm==1.4.7
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.10.1
PyMySQL==1.0.2
PyNaCl==1.1.2
pyotp==2.6.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
s3transfer==0.3.4
scp==0.13.2
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
stripe==2.49.0
system-service==0.3
systemd-python==234
tqdm==4.61.1
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.3
vine==5.0.0
wadllib==1.3.2
wcwidth==0.2.5
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
